# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "SVD Payment API",
        "contact": {},
        "version": "1.0"
    },
    "host": "103.210.35.189:8120",
    "basePath": "/",
    "paths": {
...
        "/api/methods": {
            "get": {
                "description": "List payment methods with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentMethod"
                ],
                "summary": "List payment methods",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by Payment Method ID",
                        "name": "payment_method_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by method code (partial match)",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by method name (partial match)",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentMethodListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentMethodListResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new payment method",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentMethod"
                ],
                "summary": "Create payment method",
                "parameters": [
                    {
                        "description": "Create Payment Method request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreatePaymentMethodRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentMethodDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentMethodDetailResponse"
                        }
                    }
                }
            }
        },
        "/api/methods/{id}": {
            "get": {
                "description": "Get detail of payment method by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentMethod"
                ],
                "summary": "Get payment method",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment Method ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentMethodDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentMethodDetailResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update existing payment method by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentMethod"
                ],
                "summary": "Update payment method",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment Method ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Payment Method request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdatePaymentMethodRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentMethodDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentMethodDetailResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete payment method by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PaymentMethod"
                ],
                "summary": "Delete payment method",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment Method ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentMethodDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentMethodDetailResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.CreatePaymentMethodRequest": {
            "type": "object",
            "required": [
                "code",
                "desc",
                "name",
                "order_num",
                "user_action"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 25
                },
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50
                },
                "order_num": {
                    "type": "integer"
                },
                "user_action": {
                    "type": "string",
                    "maxLength": 25
                }
            }
        },
        "models.PageMetadata": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_item": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
...
        "models.PaymentMethodDetailResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/models.PaymentMethodResponse"
                },
                "errors": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "paging": {
                    "$ref": "#/definitions/models.PageMetadata"
                }
            }
        },
        "models.PaymentMethodListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PaymentMethodResponse"
                    }
                },
                "errors": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "paging": {
                    "$ref": "#/definitions/models.PageMetadata"
                }
            }
        },
        "models.PaymentMethodResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "order_num": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_action": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdatePaymentMethodRequest": {
            "type": "object",
            "required": [
                "code",
                "desc",
                "name",
                "order_num",
                "user_action"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 25
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50
                },
                "order_num": {
                    "type": "integer"
                },
                "user_action": {
                    "type": "string",
                    "maxLength": 25
                }
            }
        }
    }
}